+++ src/NEventStore.Serialization.Rijndael.Tests/SerializerFixture.cs
@@ -9,7 +9,9 @@ namespace NEventStore.Serialization.AcceptanceTests
 
         public SerializerFixture()
         {
+#pragma warning disable CS0618 // Type or member is obsolete
             _createSerializer = () => new RijndaelSerializer(new BinarySerializer(), EncryptionKey);
+#pragma warning restore CS0618 // Type or member is obsolete
         }
     }
 }