+++ src/NEventStore/Persistence/InMemory/InMemoryPersistenceEngine.cs
@@ -186,9 +186,8 @@ namespace NEventStore.Persistence.InMemory
                 DateTime commitStamp,
                 Int64 checkpointToken,
                 IDictionary<string, object> headers,
-                IEnumerable<EventMessage> events)
-                : base(bucketId, streamId, streamRevision, commitId, commitSequence, commitStamp, checkpointToken, headers,
-                      events is IList<EventMessage> list ? list : events)
+                ICollection<EventMessage> events)
+                : base(bucketId, streamId, streamRevision, commitId, commitSequence, commitStamp, checkpointToken, headers, events)
             { }
         }